backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary		# PostgreSQL
sqlalchemy
pdfplumber
python-docx
python-multipart
requests
bs4
beautifulsoup4		# parsing html with a URL
spacy			#NLP Library
scikit-learn>=1.3.0
apscheduler
schedule		# match/ats sccore calculation
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openai
pydantic
alembic     # database migration
pydantic[email] # email validation
itsdangerous    # for creating tokens
python-docx # for creating word documents
pytest  # for testing
requests-mock # for mocking requests in tests
pandas==2.2.3
# JDI (Job Daily Intelligence) dependencies
google-auth-oauthlib     # Gmail OAuth flow
google-api-python-client # Gmail API client
cryptography             # Fernet encryption for OAuth tokens
trafilatura              # Fallback HTML→text extraction for JD pages